okay that way perfect now add a buy e coffes button. addd this text ( built by a sicle cell disodered warroior for warriors every where. Now you can show ypu over bearing anuties, parent and well wisher just why that drug isnt eant for you. note this app is at owners risk.
also add kind of notice which users must aceeot that they use as is and a mini privacy policy that we dont use or process anyof their data. this noticeshould only ever be displayed once.

added the email: olatundemarvelousanthony@gmial.com and a twitter button that actully links to the profile olamarvelcreate this would be for contacting me.

so in addition to being onthe page the buy me a coffes and y contact show lso be display as a odal when ever the user wnat to sahre and the iage is cyretly generating.